Amélioration des parapheurs (send et received)

+++ app/Http/Controllers/BatchController.php
@@ -9,7 +9,8 @@ class BatchController extends Controller
 {
     public function index()
     {
-        $mailBatches = Batch::all(); // Correction de la casse camelCase
+
+        $mailBatches = Batch::where('organisation_holder_id', auth()->user()->organisation->id)->get();
         return view('batch.index', compact('mailBatches'));
     }
 
@@ -23,16 +24,19 @@ class BatchController extends Controller
 
     public function store(Request $request)
     {
+
         $validatedData = $request->validate([
             'code' => 'nullable|unique:batches|max:10',
             'name' => 'required|max:100',
         ]);
 
+        $validatedData['organisation_holder_id'] = auth()->user()->organisation->id;
+
+
         Batch::create($validatedData);
 
         return redirect()->route('batch.index')->with('success', 'Mail batch created successfully.');
     }
-
 
 
 
@@ -61,6 +65,9 @@ class BatchController extends Controller
             'code' => 'nullable|unique:batches,code,' . $mailBatch->id . '|max:10',
             'name' => 'required|max:100',
         ]);
+
+        $validatedData = ['organisation_holder_id' => auth()->user()->organisation->id];
+
         $mailBatch->update($validatedData);
         return redirect()->route('batch.index')->with('success', 'Mail batch updated successfully.');
     }
@@ -70,8 +77,13 @@ class BatchController extends Controller
 
     public function destroy(Batch $mailBatch)
     {
-        $mailBatch->delete();
 
-        return redirect()->route('batch.index')->with('success', 'Mail batch deleted successfully.');
+        if ($mailBatch->mails->count() <= 0) {
+            $mailBatch->delete();
+            return redirect()->route('batch.index')->with('success', 'Mail batch deleted successfully.');
+        }
+        return redirect()->route('batch.index')->with('error', 'Cannot delete mail batch with associated mails.');
     }
+
+
 }
